+++ DBManager/Reapack_Header.lua
@@ -1,27 +1,17 @@
 --[[
 @metapackage true
 @description DBManager
-@version 0.81beta
+@version 0.9alpha
 @author Claudiohbsantos
 @link http://claudiohbsantos.com
-@date 2019 08 19
+@date 2019 08 21
 @about
     Database manager GUI for Reaper Media Explorer.
     Uses DBAssistant to make changes - https://github.com/Claudiohbsantos/DBAssistant . The executable will be automatically downloaded upon installation. 
     Depends on JS Reascript API and Lokasenna GUI Library v2.
 @changelog
-    # DBManager v0.81beta
-
-    ## BUGS
-
-    - error when searching without making selection
-    - false error dialogue when importing config file
-    - mismatched references  on list after save
-    - fixed preview copy destination on import sfx window
-
-    ## IMPROVEMENTS
-
-    - Single popup for new DB creation
+    - Added mac compatibility
+    - Includes DBAssistant 0.3.5
 @provides
     [main] DBManager.lua > DBManager.lua
     [nomain] lua_modules/json.lua > lua_modules/json.lua
@@ -31,6 +21,11 @@
     [nomain] reascript_modules/DBM_actions.lua > reascript_modules/DBM_actions.lua
     Documentation/* > Documentation/
     changelog.md > changelog.md
-    [windows] dbassistant.exe https://github.com/Claudiohbsantos/DBAssistant/releases/download/v0.3.4/dbassistant.exe
-    dbassistant_changelog.md https://github.com/Claudiohbsantos/DBAssistant/releases/download/v0.3.4/changelog.md
+    [windows] dbassistant.exe https://github.com/Claudiohbsantos/DBAssistant/releases/download/v0.3.5/dbassistant.exe
+    [darwin] dbassistant https://github.com/Claudiohbsantos/DBAssistant/releases/download/v0.3.5/dbassistant
+    dbassistant_changelog.md https://github.com/Claudiohbsantos/DBAssistant/releases/download/v0.3.5/changelog.md
+    [darwin] osx_launchers/osx_add.sh > osx_launchers/osx_add.sh 
+    [darwin] osx_launchers/osx_deduplicate.sh > osx_launchers/osx_deduplicate.sh 
+    [darwin] osx_launchers/osx_export.sh > osx_launchers/osx_export.sh 
+    [darwin] osx_launchers/osx_version.sh > osx_launchers/osx_version.sh 
 --]]