+++ DBManager/osx_launchers/osx_add.sh
@@ -1,0 +1,5 @@
+#! /bin/sh
+
+DIR=$(dirname $BASH_SOURCE)/..
+
+$DIR/dbassistant add $DIR/tempFile_DBAssistantFromReaper.json