+++ DBManager/osx_launchers/osx_version.sh
@@ -1,0 +1,5 @@
+#! /bin/sh
+
+DIR=$(dirname $BASH_SOURCE)/..
+
+$DIR/dbassistant -V